Created a report service (K2530341_ReportService) to generate reports like overdue books and the most borrowed books.
Made a main class called SmartLibraryManagement to start the application.

Updated the project to include new classes and files for the Maven compiler plugin.

Added managers to handle CSV files for notifications, books, and users.

Added commands for reserving and borrowing books using a command system.

Improved the book feature by adding decorators to show recommended and featured books.

Used a state system to track the status of books (whether they are available, borrowed, or reserved).